# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Three Kingdoms City Node",
    "version": "1.3.0"
  },
  "paths": {
...
    "/social/join-lord": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Join Lord",
        "operationId": "join_lord_social_join_lord_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinLordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/social/recruit": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Recruit",
        "operationId": "recruit_social_recruit_post",
        "parameters": [
          {
            "name": "lord_agent_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lord Agent Id"
            }
          },
          {
            "name": "target_agent_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Target Agent Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/social/message": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Message",
        "operationId": "message_social_message_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/social/messages": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "List Messages",
        "operationId": "list_messages_social_messages_get",
        "parameters": [
          {
            "name": "agent_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Agent Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/social/faction/create": {
      "post": {
        "tags": [
          "social"
        ],
        "summary": "Create Faction",
        "operationId": "create_faction_social_faction_create_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFactionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/social/factions": {
      "get": {
        "tags": [
          "social"
        ],
        "summary": "List Factions",
        "operationId": "list_factions_social_factions_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CreateFactionRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 64,
            "minLength": 2,
            "title": "Name"
          },
          "leader_agent_id": {
            "type": "integer",
            "title": "Leader Agent Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "leader_agent_id"
        ],
        "title": "CreateFactionRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JoinLordRequest": {
        "properties": {
          "agent_id": {
            "type": "integer",
            "title": "Agent Id"
          },
          "lord_agent_id": {
            "type": "integer",
            "title": "Lord Agent Id"
          }
        },
        "type": "object",
        "required": [
          "agent_id",
          "lord_agent_id"
        ],
        "title": "JoinLordRequest"
      },
...
      "MessageRequest": {
        "properties": {
          "from_agent_id": {
            "type": "integer",
            "title": "From Agent Id"
          },
          "to_agent_id": {
            "type": "integer",
            "title": "To Agent Id"
          },
          "message_type": {
            "type": "string",
            "title": "Message Type"
          },
          "content": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Content"
          }
        },
        "type": "object",
        "required": [
          "from_agent_id",
          "to_agent_id",
          "message_type",
          "content"
        ],
        "title": "MessageRequest"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}